savepoint

+++ src/efc/test/testhelpers/literaltokenstream.cpp
@@ -6,7 +6,7 @@ LiteralTokenStream::LiteralTokenStream(const vector<Parser::token_type>& tokens)
   m_front(0),
   m_size(tokens.size()),
   m_stream(new Parser::symbol_type[m_size]) {
-  for (int i=0; i<tokens.size(); ++i) {
+  for (size_t i=0; i<tokens.size(); ++i) {
     auto tmp = ParserApiExt::makeToken(tokens[i]);
     m_stream[i].move(tmp);
   }
